haskell: disable warnings on unused imports and variables
Since we are running with -Werror (maybe we shouldn't) and using cpp to
configure whether arguments are used or not depending on the arch (i.e.
ARM vs ARM_HYP), it makes no sense to make unused variables and imports
an error.

Open to ideas on where to go from here.