remove lemma hv_inv_ex from Syscall_AI assumptions
The lemma is not true on ARM_HYP. It is also not needed for AInvs, although
it looks like it may be used in Refine/CRefine.

It'll have to be replaced with a set of more specific assumptions, which
should be added on master, so I'm just removing it on this branch for now.